# swagger.yaml
openapi: 3.0.3
info:
  title: Cerberus License API
  description: |
    Cerberus 软件授权管理系统 API 文档

    ## 认证方式
    - 公开接口：无需认证
    - 管理接口：需要在 Header 中携带 `Authorization: Bearer <token>`

    ## 统一响应格式
    所有接口返回统一的 JSON 格式：
    ```json
    {
      "code": 0,
      "message": "ok",
      "data": {...}
    }
    ```
    - `code`: 0 表示成功，非 0 表示错误
    - `message`: 响应消息
    - `data`: 响应数据
  version: 1.0.0
  contact:
    name: Cerberus Team
servers:
  - url: http://localhost:8080
    description: 本地开发服务器
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT Token 认证
  schemas:
...
    APIResponse:
      type: object
      properties:
        code:
          type: integer
          description: 业务状态码，0 表示成功
          example: 0
        message:
          type: string
          description: 响应消息
          example: "ok"
        data:
          type: object
          description: 响应数据

    # ==================== License 相关 ====================
...
    Machine:
      type: object
      properties:
        id:
          type: string
          description: 机器唯一标识符（UUID）
          example: "660e8400-e29b-41d4-a716-446655440000"
        license_id:
          type: string
          description: 所属 License ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        fingerprint:
          type: string
          description: 机器指纹（硬件信息哈希）
          example: "abc123def456..."
        hostname:
          type: string
          description: 主机名
          example: "MyPC"
        os:
          type: string
          description: 操作系统
          example: "windows"
        arch:
          type: string
          description: 系统架构
          example: "amd64"
        ip:
          type: string
          description: 当前 IP 地址
          example: "192.168.1.100"
        ip_binding:
          type: string
          description: 绑定的 IP 地址
          example: "192.168.1.100"
        last_seen:
          type: string
          format: date-time
          description: 最后活跃时间
        status:
          type: string
          description: 机器状态
          enum: [active, revoked]
          example: "active"
        geo_country:
          type: string
          description: 地理位置国家代码
          example: "CN"
        geo_region:
          type: string
          description: 地理位置地区代码
          example: "Beijing"
        geo_city:
          type: string
          description: 地理位置城市
          example: "Beijing"
        geo_latitude:
          type: number
          format: double
          description: 纬度
          example: 39.9042
        geo_longitude:
          type: number
          format: double
          description: 经度
          example: 116.4074
        created_at:
          type: string
          format: date-time
          description: 创建时间
    AuditLog:
      type: object
      properties:
        id:
          type: string
          description: 审计日志 ID
        license_id:
          type: string
          description: 关联的 License ID
        action:
          type: string
          description: 操作类型
          example: "activate"
        detail:
          type: string
          description: 操作详情
          example: "machine activated: abc123..."
        ip:
          type: string
          description: 操作来源 IP
          example: "192.168.1.100"
        created_at:
          type: string
          format: date-time
          description: 创建时间

    # ==================== 请求体 ====================
...
    AuditLogsRequest:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: License ID
          example: "550e8400-e29b-41d4-a716-446655440000"
    RevokeMachineRequest:
      type: object
      required:
        - license_id
        - machine_id
      properties:
        license_id:
          type: string
          description: License ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        machine_id:
          type: string
          description: 机器 ID
          example: "660e8400-e29b-41d4-a716-446655440000"

    # ==================== 公开接口请求 ====================
    ActivateRequest:
      type: object
      required:
        - license_id
        - fingerprint
      properties:
        license_id:
          type: string
          description: License ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        fingerprint:
          type: string
          description: 机器指纹
          example: "abc123def456..."
        hostname:
          type: string
          description: 主机名
          example: "MyPC"
        os:
          type: string
          description: 操作系统
          example: "windows"
        arch:
          type: string
          description: 系统架构
          example: "amd64"
        ip:
          type: string
          description: IP 地址
          example: "192.168.1.100"
...
    HeartbeatRequest:
      type: object
      required:
        - license_id
        - fingerprint
      properties:
        license_id:
          type: string
          description: License ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        fingerprint:
          type: string
          description: 机器指纹
          example: "abc123def456..."
        ip:
          type: string
          description: IP 地址
          example: "192.168.1.100"
...
paths:
...
  /api/v1/activate:
    post:
      tags:
        - Public
      summary: 激活机器
      description: |
        客户端使用此接口激活机器绑定。

        **流程：**
        1. 验证 License 是否有效
        2. 检查机器绑定数量是否已达上限
        3. 绑定当前机器
        4. 返回激活结果（包含地理位置告警信息）
      operationId: activate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateRequest'
      responses:
        '200':
          description: 激活成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          machine:
                            $ref: '#/components/schemas/Machine'
                          geoip_alert:
                            type: string
                            description: 地理位置告警信息
                            example: "remote login detected (800 km away)"
                          is_new_machine:
                            type: boolean
                            description: 是否为新机器
        '400':
          description: 激活失败（License 无效、机器数量超限等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
...
  /api/v1/heartbeat:
    post:
      tags:
        - Public
      summary: 心跳上报
      description: 客户端定期上报心跳，更新机器最后活跃时间
      operationId: heartbeat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeartbeatRequest'
      responses:
        '200':
          description: 心跳上报成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
        '400':
          description: 心跳上报失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
...
  /api/v1/licenses/audit:
    post:
      tags:
        - License
      summary: 获取审计日志
      description: 获取指定 License 的所有审计日志
      operationId: getAuditLogs
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditLogsRequest'
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AuditLog'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  # ==================== 机器管理接口 ====================
  /api/v1/licenses/machines/revoke:
    post:
      tags:
        - Machine
      summary: 吊销机器
      description: |
        管理员主动吊销单个机器绑定，让用户能够绑定新机器。
        与客户端自助换绑不同，此操作不消耗换绑次数。
      operationId: revokeMachine
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeMachineRequest'
      responses:
        '200':
          description: 吊销成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "machine revoked, you can now activate a new device"
        '400':
          description: 吊销失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'